refactor(deck_results_list): split DeckResultsList into mixin package
Move widgets/lists/deck_results_list.py into a package: frame.py holds
the class definition and __init__; properties.py owns the count/string
accessors and pure emoji/line splitters; handlers.py owns the append,
clear, draw, measure, and font-fitting routines.
