+++ widgets/lists/deck_results_list/__init__.py
@@ -1,0 +1,7 @@
+"""Deck results list widget package."""
+
+from __future__ import annotations
+
+from widgets.lists.deck_results_list.frame import DeckResultsList
+
+__all__ = ["DeckResultsList"]
